yayasan_abc/requirements_yayasan.txt
grpcio==1.68.0
grpcio-tools==1.68.0
django-grpc-framework==0.1.0
